requirements.txt
aiogram==2.11.2
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
emoji==1.1.0
et==0.0.2
idna==2.10
multidict==5.1.0
pkg-resources==0.0.0
pyTelegramBotAPI==3.7.6
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
typing-extensions==3.7.4.3
ujson==4.0.1
urllib3==1.26.3
uvloop==0.14.0
yarl==1.6.3
